Struct jellyfish_merkle::iterator::JellyfishMerkleIterator

pub struct JellyfishMerkleIterator<R: TreeReader> { /* fields omitted */ }

The JellyfishMerkleIterator implementation.

Methods

impl<R> JellyfishMerkleIterator<R> where
R: TreeReader,

pub fn new(
reader: Arc<R>,
version: Version,
starting_key: HashValue
) -> Result<Self>

Constructs a new iterator. This puts the internal state in the correct position, so the following next call will yield the smallest key that is greater or equal to starting_key.